software: move the NV12 consumer seam to a backend-neutral header
INv12Consumer and MakeNv12Consumer lived in encoder_sink.h, which pulls in the
software backend's ISurfaceSink. Split the seam into nv12_consumer.h so a second
producer -- the coming headless-EGL backend, which packs NV12 on the GPU -- can
feed the same file/webrtc consumers without depending on the CPU sink. No
behavior change; encoder_sink.h and the consumer TUs just include the new header.
